// src/types/instance.ts
export type Instance = {
  id: number;
  name: string;
  description: string;
  billing_cycle: string;
  created_at_for_human: string;
  status: number;
  private_ip: string;
  public_ip: string;
  port: string;
  client_token: string;
  code: string;
  region: string;
  server_region: string;
  expiry_date: string;
  price_for_human: string;
  price: string;
  status_in_string: string;
};